+++ .vscode/settings.json
@@ -7,6 +7,7 @@
     "alconna",
     "arithmatex",
     "asgi",
+    "Behaviour",
     "behaviours",
     "betterem",
     "cheatsheet",
@@ -17,6 +18,7 @@
     "docformat",
     "favicon",
     "fontawesome",
+    "HRTX",
     "inlinehilite",
     "isawaitable",
     "Kugou",
@@ -26,12 +28,14 @@
     "linenums",
     "Migu",
     "miraijvm",
+    "MISRO",
     "mkdocstrings",
     "pycon",
     "pymdownx",
     "pypi",
     "quickstart",
     "smartsymbols",
+    "sourcery",
     "starlette",
     "superfences",
     "tasklist",
